refactor: separate Analyze and cli components
